ItemNo 형변환(int -> long)

+++ src/main/java/kr/co/_29cm/homework/backend/model/entity/Item.java
@@ -17,7 +17,7 @@ public class Item {
 
     @Id
     @Column(name = "ITEM_NO")
-    private int itemNo;          //상품번호
+    private long itemNo;          //상품번호
 
     @Column(name = "NAME")
     private String name;            //상품명
